Split out sub widgets, set up animated containers
